Added Actor framework
Actors are game objects that have abilities and properties such as health. Actors can be influenced by applying effects.
The ActorEffect class represents the description of the effect to be applied. For now there are only Damage, Kill and Heal effects, but more can be added. Additional properties related to the process of applying the effect can be added as well.