Refactoring of Player and Enemy classes
Made separation between character and character controller.
Added Character class which extends Actor and handles generic character abilities.
Added EnemyCharacter and PlayerCharacter classes which take care of specific character properties, such as animations or sounds.
Added PlayerController and EnemyAI, which serve the purpose of driving Characters.
With this, it is now possible to mix and match characters with controllers, making it easier to add other players, or different AIs.