Fixed the background parallax
The old parallax was moving the backgrounds in the wrong direction and it was somewhat more confusing.
The backgrounds should not move opposite the camera movement, but along side it with fraction of its displacement.
That way those backgrounds that are at a greater distance will move closely with the camera, appearing to stay still in the distance, while those closer (smaller fraction) will appear to move faster. And the closest one (foreground) is naturally the fastest because it doesn't follow the camera at all.