Added two new features to RangedWeapon
RangedWeapon now looks for a 'muzzle' transform in its hierarchy, which denotes the exit point of the projectile.
Added reloadTime, after the ranged weapon is Fired, it can't be fired again until the amount of time in reloadTime has passed.
Also, the PlayerCharacter now rotates his ranged weapon based on orientation in the range from 45 degrees up to 45 degrees down.
A Hovering Character whose Rigidbody has no linear drag, will now stop in the air after half a second if there is no force applied to it (it will no longer drift endlessly).
The attack interval of the flaying saucers is now configurable.