Changed the way enemies deal contact damage
Added minDamageInterval to EnemyCharacter. Enemies now only deal contact damage if they didn't do so for a while.
Fixed few bugs where triggers and collisions were reported twice.
The Cars and Buses on the background are now attached to the 'Bank' background object. This way, they work correctly with the parallax.