Refactored LayBombs. Made some reorganization.
LayBombs has been refactored into BombContainer which only keeps track of a number of bombs and their type.
The action of actually dropping a bomb has been moved to Character. As long as a character has a BombContainer and has bombs in it, it can use the drop bomb mechanic.
The Bomb class is no longer responsible for counting bombs or preventing the user from dropping them. It only knows how to explode. (Though for now it is still responsible for triggering the spawning of next Pickup).
Moved the PlayerController, EnemyAI and FlyingEnemyAI to Actor/Controller/ folder.
Moved the BombPickup and HealthPickup to Pickups/ folder.