Added crosshair, refactored some of the UI
Renamed HUD/ to UI/
Removed Follow Player.
Added UIHealthbar, which can be placed on or assigned an actor and it will display his health.
Added UICrosshair, which when placed on a character displays that character's orientation.
Added UIBomb, which takes care of showing and hiding the bomb icon, this is no longer a responsibilty of the PlayerCharacter.